// src/types/API.ts
export type CreateTodoInput = {
  id?: string | null
  name: string
  description?: string | null
}
// ...
export type UpdateTodoInput = {
  id: string
  name?: string | null
  description?: string | null
}

export type DeleteTodoInput = {
  id?: string | null
}